refactor: move init modules and more business logic to vm. added destroy method for clean up. wrote getLabel method
